Dispatching the Message
Phew! Status check: we have a message class, we have a handler class, and thanks to some smartness from Symfony, Messenger knows these are linked together. The last thing we need to do is... actually send the command, or "message", to the bus!

Head over to ImagePostController. This is the endpoint that uploads our image and adds Ponka to it. Fetch the message bus by adding a new argument with the MessageBusInterface type-hint.

Then... right above all the Ponka image code - we'll leave all of that there for the moment - say $message = new AddPonkaToImage(). And then $messageBus->dispatch($message).

That's it! dispatch() is the only method on that object... it doesn't get any more complicated than this.

So... let's try it! If everything works, this AddPonkaToImage object should be passed to __invoke() and then we'll dump it. Since this will all happen on an AJAX request, we'll use a trick in the profiler to see if it worked.

Head back and refresh the page... just to be safe. Upload a new photo and... when it finishes, down on the web debug toolbar, hover over the arrow icon to find... nice! Here is that AJAX request. I'll hold Command and click the link to open it in a new tab. This is the profiler for that AJAX request. Click the "Debug" link on the left.

Ha! There it is! This shows us that our dump() code was executed during the AJAX request! It worked! We pass the message to the message bus and then it calls the handler.

Of course... our handler doesn't do anything yet. Next, let's move all of the Ponkafication logic from our controller into the handler.