Inside our controller, after we save the new file to the filesystem, we're creating a new AddPonkaToImage object and dispatching it to the message bus... or technically the "command" bus... because we're currently using it as a command bus. The end result is that the bus calls the __invoke() method on our handler and passes it that object. Messenger understands the connection between the message object and handler thanks to the argument type-hint and this interface.
Command Bus: Beautifully Disappointing

By the way, you might be thinking:

Wait... the whole point of a "command" bus is to... just "call" this __invoke() method for me? Couldn't I just... ya know... call it myself and skip a layer?

And... yes! It's that simple! It should feel completely underwhelming at first!

But having that "layer", the "bus", in the middle gives us two nice things. First, out code is more decoupled: the code that creates the "command" - our controller in this case - doesn't know or care about our handler. It dispatches the message and moves on. And second, this simple change is going to allow us to execute handlers asynchronously. More on that soon.

Moving code into the Handler

Back to work: all the code to add Ponka to the image is still done inside our controller: this gets an updated version of the image with Ponka inside, another service actually saves the new image onto the filesystem, and this last bit - $imagePost->markAsPonkaAdded() - updates a date field on the entity. It's only a few lines of code... but that's a lot of work!

Copy all of this, remove it, and I'll take my comments out too. Paste all of that into the handler. Ok, no surprise, we have some undefined variables. $ponkaficator, $photoManager and $entityManager are all services. In the controller... on top, we were autowiring those services into the controller method. We don't need $ponkaficator anymore.

Anyways, how can we get those services in our handler? Here's the really cool thing: the "message" class - AddPonkaToImage is a simple, "model" class. It's kind of like an entity: it doesn't live in the container and we don't autowire it into our classes. If we need an AddPonkaToImage object, we say: new AddPonkaToImage(). If we decide to give that class any constructor arguments - more on that soon - we pass them right here.

But the handler classes are services. And that means we can use, good, old-fashioned dependency injection to get any services we need.

Add public function __construct() with, let's see here, PhotoPonkaficator $ponkaficator, PhotoFileManager $photoManager and... we need the entity manager: EntityManagerInterface $entityManager.

I'll hit Alt + Enter and select Initialize Fields to create those properties and set them.

Now... let's use them: $this->ponkaficator, $this->photoManager, $this->photoManager again... and $this->entityManager.