Message Class Data
Nice! This leaves us with just one undefined variable: the actual $imagePost that we need to add Ponka to. Let's see... in the controller, we create this ImagePost entity object... which is pretty simple: it holds the filename on the filesystem... and a few other minor pieces of info. This is what we store in the database.

Back in AddPonkaToImageHandler, at a high level, this class needs to know which ImagePost it's supposed to be working on. How can we pass that information from the controller to the handler? By putting it on the message class! Remember, this is our class, and it can hold whatever data we want.

So now that we've discovered that our handler needs the ImagePost object, add a public function __construct() with one argument: ImagePost $imagePost. I'll do my usual Alt+Enter and select "Initialize fields" to create and set that property. Down below, we'll need a way to read that property. Add a getter: public function getImagePost() with an ImagePost return type. Inside, return $this->imagePost.

And really... you can make this class look however you want: we could have made this a public property with no need for a constructor or getter. Or you could replace the constructor with a setImagePost(). This is the way I like to do it... but it doesn't matter: as long as it holds the data you want to pass to the handler... you're good!

Anyways, now we're dangerous! Back in ImagePostController, down here, AddPonkaToImage now needs an argument. Pass it $imagePost. Then, over in the handler, finish this with $imagePost = $addPonkaToImage->getImagePost().

I love it! So that's the power of the message class: it really is like you're writing a message to someone that says:

I want you to do a task and here's all the information that you need to know to do that task.

Then, you hand that off to the message bus, it calls the handler, and the handler has all the info it needs to do that work. It's a simple... but really neat idea.

Let's make sure it all works: move over and refresh just to be safe. Upload a new image and... it still works!

Next: there's already one other job we can move to a command-handler system: deleting an image.