To find all the handlers when run "php bin/console debug:messenger" in Terminal, Symfony looks in the src/ directory for classes that implement MessageHandlerInterface. And... I forgot that part! Add implements MessageHandlerInterface
Run debug:messenger again:
php bin/console debug:messenger

Now it sees it! Let's try it again: close up the profiler, try hitting "x" and... this time it works!

Status report: we have two messages and each has a handler that's potentially doing some pretty heavy work, like image manipulation or talking across a network if files are stored in the cloud. It's time to talk about transports: the key concept behind taking this work and doing it asynchronously so that our users don't have to wait for all that heavy work to finish before getting a response.