Suppose you need your friend to come over and watch your dog for the weekend - let's call her Molly. So you write them a message explaining all the details they need to know: how often to feed Molly, when to walk her, exactly where she likes to be scratched behind the ears, your favorite superhero movie and the name of your childhood best friend. Wait... those last two things... while fascinating... have nothing to do with watching your dog Molly!
And this touches on a best-practice for designing your message classes: make them contain all the details the handler needs... and nothing extra. This isn't an absolute rule... it just makes them leaner, smaller and more directed.

Passing the Entity Id

If you think about our message, we don't really need the entire ImagePost object. The smallest thing that we could pass is actually the id... which we could then use to query for the ImagePost object and get the filename.

Change the constructor argument to int $imagePostId. I'll change that below and go to Code -> Refactor to rename the property. Oh, and brilliant! It also renamed my getter to getImagePostId(). Update the return type to be an int. We can remove the old use statement as extra credit. Next, in ImagePostController, search for AddPonkaToImage and... change this to $imagePost->getId(). Our message class is now as small as it can get. Of course, this means that we have a little bit extra work to do in our handler. First, the $imagePost variable is not... well.. an ImagePost anymore! Rename it to $imagePostId. To query for the actual object, add a new constructor argument: ImagePostRepository $imagePostRepository. I'll hit Alt + Enter -> Initialize Fields to create that property and set it. Back in the method, we can say $imagePost = $this->imagePostRepository->find($imagePostId).

That's it! And this fixes our Doctrine problem! Now that we're querying for the entity, when we call flush(), it will correctly save it with an UPDATE. We can remove the persist() call because it's not needed for updates.

Let's try it! Because we just changed code in our handler, hit Ctrl+C to stop our worker and then restart it:

 php bin/console messenger:consume -vv
Here we go! Upload a new file... check the worker - yep, it processed just fine - and... refresh! Yes! No duplication, Ponka is visiting my workshop and the date is set!